Added proposal for config files
One problem with controller scripts is,, that one needs to create a full
copy with it's own UUID for any changes or settings. So as a temporary
measure I suggest having a config file for each controller script
containing the same variables as the config part on top of each
controller script, overriding them if present/different. Ideally those
would reside in the user folder and wouldn't get overridden by
installations.
In the long run, a way to have such options in the GUI would be
preferable of course... ;-)